# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "รวมรวม API สำหรับ Application  ที่ใช้ในโครงการ [http://mmc_app1.mrta.co.th/](http://mmc_app1.mrta.co.th/) ",
        "version": "1.0.1",
        "title": "MRTA API DOCUMENTS"
    },
    "host": "mmc_app1.mrta.co.th",
    "basePath": "/api/v2",
    "schemes": [
        "http"
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces" : [
        "application/json"
    ],
    "paths": {
...
        "/stops/{agency_key}/{route_id}": {
            "parameters": [
                {
                    "name": "agency_key",
                    "in": "path",
                    "required": true,
                    "description": "ค้นหา stops ด้วย agency_key",
                    "type": "string"
                },
                {
                    "name": "route_id",
                    "in": "path",
                    "required": true,
                    "description": "ค้นหา stops ด้วย route_id",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Stops"
                ],
                "summary": "รายการ Stops ทั้งหมด",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stop"
                        }
                    }
                }
            }
        },
        "/stops/{agency_key}": {
            "parameters": [
                {
                    "name": "agency_key",
                    "in": "path",
                    "required": true,
                    "description": "ค้นหา stops ด้วย agency_key",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Stops"
                ],
                "summary": "รายการ Stops ทั้งหมด",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stop"
                        }
                    }
                }
            }
        },
...
        "/stoptimes/{agency_key}/{route_id}": {
            "parameters": [
                {
                    "name": "agency_key",
                    "in": "path",
                    "required": true,
                    "description": "ค้นหา stoptims ด้วย agency_key",
                    "type": "string"
                },
                {
                    "name": "route_id",
                    "in": "path",
                    "required": true,
                    "description": "ค้นหา stoptims ด้วย route_id",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "StopTimes"
                ],
                "summary": "รายการ Trips ทั้งหมด",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stop_time"
                        }
                    }
                }
            }
        },
...
        "/routeinfowithtrip/{trip_id}": {
            "parameters": [
                {
                    "name": "trip_id",
                    "in": "path",
                    "required": true,
                    "description": "ค้นหา route info ด้วย trip id",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "RouteInfo"
                ],
                "summary": "รายการ Route Info ทั้งหมด",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routeinfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "stop": {
            "type": "object",
            "properties": {
                "agency_key": {
                    "type": "string"
                },
                "stop_id": {
                    "type": "string"
                },
                "stop_code": {
                    "type": "string"
                },
                "stop_name": {
                    "type": "string"
                },
                "stop_desc": {
                    "type": "string"
                },
                "stop_lat": {
                    "type": "string"
                },
                "stop_lon": {
                    "type": "string"
                },
                "zone_id": {
                    "type": "string"
                },
                "stop_url": {
                    "type": "string"
                },
                "location_type": {
                    "type": "string"
                },
                "parent_station": {
                    "type": "string"
                },
                "stop_timezone": {
                    "type": "string"
                },
                "wheelchair_boarding": {
                    "type": "number"
                },
                "level_id": {
                    "type": "number"
                },
                "platform_code": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                }
            }
        },
...
        "stop_time": {
            "type": "object",
            "properties": {
                "agency_key": {
                    "type": "String"
                  },
                  "trip_id": {
                    "type": "String"
                  },
                  "arrival_time": {
                    "type": "String"
                  },
                  "departure_time": {
                    "type": "String"
                  },
                  "stop_id": {
                    "type": "String"
                  },
                  "stop_sequence": {
                    "type": "Number"
                  },
                  "stop_headsign": {
                      "type": "String"
                  },
                  "pickup_type": {
                    "type": "Number"
                  },
                  "drop_off_type": {
                    "type": "Number"
                  },
                  "shape_dist_traveled": {
                    "type": "Number"
                  },
                  "timepoint": {
                    "type": "Number"
                  }
            }
        },
...
        "routeinfo": {
            "type": "object",
            "properties": {
                "agency_key": {
                    "type": "String"
                },
                "feed_publisher_name": {
                    "type": "String"
                },
                "feed_publisher_url": {
                    "type": "String"
                },
                "feed_lang": {
                    "type": "String"
                },
                "feed_start_date": {
                    "type": "Number"
                },
                "feed_end_date": {
                    "type": "Number"
                },
                "feed_version": {
                    "type": "String"
                }
            }
        },
...
    }
}